function_upload_data_into_bigquery/requirements.txt
google-cloud-bigquery
google-cloud-storage
functions-framework
pytest